Add optimization flag reorganize texture assets

+++ include/cub3d_types.h
@@ -51,7 +51,7 @@ typedef struct s_player
 	int		turn_dir; /* 1 for right, -1 for left */
 	int		strafe_dir;
 	int		move_dir;
-	int		move_speed;
+	double	move_speed;
 	double	turn_spd;
 }	t_player;
